Adding a day to the auth token expiration date.
The previous code would set the expiration date of the auth token to the next day.
This was done regardless of what the stored expiration date was. This
resulted in the resizer account getting locked out when there was no activity over more than a day.

This change adds a day to the existing expiration date instead of forcing it to the next day. This allows accounts with a longer token life to continue to work even over a period of inactivity.